Deplying multiple containers to ECS
Using a similar process before I deploy this app, along with a mongdb
container to the same task in an ECS service.
Because they are in the same service, they are hosted together and you
can substitute localhost for the mongodb url instead of the container
name.

This setup utilizes a load balancer to direct requests and lets us use
the load balance DNS instead of needing to change ip addresses every
deployment.

This setup comes with a problem though, the backend deployment fails unless
you account for it needing to connect to the DB.
This can be handled with ECS container dependencies, but here has been
done with a set timeout.

This also presents problems with data persistence. For for I created a
EFS volume to achieve this, but this comes with rolling deployment issues.
Next I will replace this with mongodb Atlas