Backend updated to work with MongoDB Atlas
To avoid issues to managing DB, which given this is not a large scale
application is unneccessary, switched out to managed cloud db provider.

This is remarkably simple given the capabilities it gives, only requiring
a change to the url connection string, and the env variables that are
passed in.

This also means we can use the same DB/env for production/ development
but simply add a different db table for the two.