Added frontend to ECS deployment
ECS deployment is now a complete app accessible with a React frontend.

To add the frontend was as simple as adding another ECS task with the
React container and pointing a load balancer towards it.
This can likely be optimized to just one load balancer in future.

The main thing to understand here is Reacts' build stage, a production
React app is run in a browser, which does not have a node environment,
and therefore won't understand things like JSX or have direct access to
your server.
The build stage builds an optimied HTML/ code bundle which can itself
be hosted on a server, which will then call to your backend server when
it needs data.

Another new learning, React env variables require REACT_APP_ in front of
them otherwise they are automatically ignored to preserve potential
key/ sensitive info leaks.