Kubernetes networking
This new directory contains a fully working full stack application (
albeit without a database ) that works by making use of kubernetes
networking.

Each of the 3 apis and the frontend exist on their own deployments
and services, making necessary communications using the cluster internal
domain name. Originally these used the IP addresses, and later built in
environment names, but this approach is the most intuitive.

I set the apis to be internal access only, with only the frontend
externally accessible, as this made the most sense to be for production.
this can be easily changed by changing the service spec types to
LoadBalancer.

In order to run this application you need to create a docker image for
each of the containers, in this case auth, frontent, tasks and users,
push to dockerhub, and then include their names in the service config
files.
You can then use the kubectl apply command with each of the following:
  -f=kubernetes/auth-service.yaml
  -f=kubernetes/auth-deployment.yaml
  -f=kubernetes/frontend-service.yaml
  -f=kubernetes/frontend-deployment.yaml
  -f=kubernetes/tasks-service.yaml
  -f=kubernetes/tasks-deployment.yaml
  -f=kubernetes/users-service.yaml
  -f=kubernetes/users-deployment.yaml

You can then use 'minikube service frontend-service' to launch the react
app and interact.